Split certifier out of the queue consumer
