+++ backend/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "start": "node src/main.js | bunyan",
     "start:consumer": "node src/consumer.js",
+    "start:certifier": "node src/certifier.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint --ignore-path .gitignore ./src/",
     "lint:js:cached": "eslint --cache --ignore-path .gitignore ./src/",
